move the code to transforms.h/cpp
