+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/Transforms.h
@@ -118,4 +118,64 @@ FailureOr<std::pair<Value, Value>> rewriteFft(Operation *op, Value operand,
                                               int64_t fftLength,
                                               PatternRewriter &rewriter);
 
+/// Apply transformation to split a linalg.generic argmax reduction
+/// into a two-stage reduction using an additional parallel dimension.
+/// The transformation first computes a partial argmax over tiles (parallel),
+/// then reduces those results into a final result (reduction).
+///
+/// This pattern is specialized for reductions that yield both the maximum
+/// value and its index, using the combination of `arith.maximumf`,
+/// `arith.cmpf`, and `arith.select` ops. It assumes a known structure of the
+/// region and injects index computations to track global indices.
+///
+/// Returns the resulting partial and final linalg.generic ops, or failure
+/// if the pattern does not match or cannot be split.
+///
+/// Example:
+/// ```
+/// // Original argmax op reducing over dim=512
+/// %4:2 = linalg.generic {
+///   indexing_maps = [...],
+///   iterator_types = ["parallel", "reduction"]
+/// } ins(%arg0 : tensor<?x512xbf16>)
+///   outs(%out_val, %out_idx : tensor<?xbf16>, tensor<?xi64>) {
+/// ^bb0(%in: bf16, %out: bf16, %out_0: i64):
+///   %idx = linalg.index 1 : index
+///   %cast = arith.index_cast %idx : index to i64
+///   %max = arith.maximumf %in, %out : bf16
+///   %cmp = arith.cmpf ogt, %in, %out : bf16
+///   %sel = arith.select %cmp, %cast, %out_0 : i64
+///   linalg.yield %max, %sel : bf16, i64
+/// } -> (tensor<?xbf16>, tensor<?xi64>)
+/// ```
+/// To:
+/// ```
+/// // After splitting K=512 into 4 x 128
+/// %expanded = tensor.expand_shape %arg0 [[0], [1, 2]] : tensor<?x512xbf16>
+///     into tensor<?x4x128xbf16>
+/// %init_val = linalg.fill ins(%cst : bf16) outs(%empty : tensor<?x4xbf16>)
+/// %init_idx = linalg.fill ins(%zero : i64) outs(%empty : tensor<?x4xi64>)
+/// %partial:2 = linalg.generic {
+///   indexing_maps = [...],
+///   iterator_types = ["parallel", "reduction"]
+/// } ins(%expanded : tensor<?x4x128xbf16>)
+///   outs(%init_val, %init_idx : tensor<?x4xbf16>, tensor<?x4xi64>) {
+///   // compute global index: outer_idx * 128 + inner_idx
+///   ...
+/// }
+///
+/// // Final argmax over the tile dimension (dim=1 of ?x4)
+/// %final:2 = linalg.generic {
+///   indexing_maps = [...],
+///   iterator_types = ["reduction"]
+/// } ins(%partial#0, %partial#1)
+///   outs(%out_val, %out_idx : tensor<?xbf16>, tensor<?xi64>) {
+///   // same combiner: maximumf, cmpf, select
+///   ...
+/// }
+/// ```
+FailureOr<linalg::SplitReductionResult>
+splitArgmaxReduction(RewriterBase &rewriter, linalg::GenericOp genericOp,
+                     linalg::ControlSplitReductionFn controlSplitReductionFn);
+
 }; // namespace mlir::iree_compiler::IREE::LinalgExt
